Made fs_perms_t and fs_ftype_t 16bit types

+++ dev/diskimage/kpi/types.h
@@ -53,8 +53,8 @@ typedef int             nlink_t;
 typedef int             dev_t;
 typedef long long       off_t;
 
-typedef unsigned int    fs_ftype_t;
-typedef unsigned int    fs_perms_t;
+typedef unsigned short  fs_ftype_t;
+typedef unsigned short  fs_perms_t;
 
 
 // Disk sector address based in cylinder, head, sector notation
